platformtheme : unload protection
This prevents a recent (?) regression where opening a KDE file dialog in
a "pure Qt" app (= not linking itself to KF5KIOWidgets) would cause a
crash on exit, apparently because the QRegExpEngine globalEngineCache
contained references to objects belonging to KF5KIOWidgets (which would
already be unloaded).
The unload protection postpones the unloading of the platform theme
plugin, and thus of any libraries on which it depends.

Committed from host : Portia.local